Move GraphDataType trait from diskann-providers to diskann-disk
- Move trait definition to diskann-disk/src/data_model/graph_data_types.rs
- Move test concrete types (3 used) to diskann-disk/src/test_utils.rs,
  remove 6 unused types
- Delete graph_data_types.rs, graph_data_type_utils.rs, and empty
  traits/ module from diskann-providers
- Update all imports across diskann-disk, diskann-tools, diskann-benchmark
- Clean up stale AdjacencyListType doc comments
- Fix benchmark import brace structure
