+++ diskann-disk/src/data_model/mod.rs
@@ -15,4 +15,7 @@ pub use graph_header::GraphHeader;
 mod cache;
 pub use cache::{Cache, CachingStrategy};
 
+pub mod graph_data_types;
+pub use graph_data_types::{AdHoc, GraphDataType};
+
 pub const FP_VECTOR_MEM_ALIGN: usize = 32;
