+++ diskann-providers/src/model/graph/mod.rs
@@ -4,5 +4,3 @@
  */
 
 pub mod provider;
-
-pub mod traits;
