--- diskann-providers/src/model/graph/traits/mod.rs
@@ -1,7 +1,0 @@
-/*
- * Copyright (c) Microsoft Corporation.
- * Licensed under the MIT license.
- */
-
-pub mod graph_data_types;
-pub use graph_data_types::{AdHoc, GraphDataType};
