+++ diskann-providers/src/test_utils/mod.rs
@@ -2,12 +2,6 @@
  * Copyright (c) Microsoft Corporation.
  * Licensed under the MIT license.
  */
-pub mod graph_data_type_utils;
-pub use graph_data_type_utils::{
-    GraphDataF32VectorU32Data, GraphDataF32VectorUnitData, GraphDataF32WithU64IdVectorU32Data,
-    GraphDataF32WithU64IdVectorUnitData, GraphDataHalfByteArrayData, GraphDataI8VectorUnitData,
-    GraphDataMinMaxVectorUnitData, GraphDataU8VectorU32AssociatedData, GraphDataU8VectorUnitData,
-};
 
 mod search_utils;
 #[cfg(test)]
